# Use ASCIISet to lower execution time by ~10 to 30%