--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/iancoleman/strcase
 
 go 1.16
+
+require github.com/elliotwutingfeng/asciiset v0.0.0-20230318033524-4982e1c19ba9